refactor(lib): extract lib/utils.php (ADR-004 foundation)
Closes #916 (to_int/to_str dedupe). Sub-task of #907.

8 functions moved from lib.php to lib/utils.php: e, to_int, to_str,
format_bytes, q_int, ipam_normalise_version, base64url_encode,
base64url_decode. Pure file move; no callsite changes (functions stay
in global namespace per ADR-004 Option E).

#916 dedupe: to_str previously had two byte-identical definitions —
one in init.php (line 155, used by early HTTPS-redirect code) and one
in lib.php (line 643, guarded by function_exists for api.php/status.php
which bootstrap lib.php directly). Both removed; canonical definition
now lives in lib/utils.php. init.php require_once's lib/utils.php
immediately after config.php loads, before the first to_str() call.
lib.php require_once's it from the top so test bootstrap (which loads
lib.php directly) keeps working unchanged.

Behavioural parity test added per #1045 umbrella rule
(tests/Unit/UtilsExtractionParityTest.php) — asserts the eight
functions exist and report lib/utils.php as their declaring file via
ReflectionFunction.

Local gate: php -l (3/3 pass), phpstan L9 (0 errors), phpcs (clean),
phpunit (1231/1231 pass), lib-module-linter (exit 0).

Refs #1221.
